# nodejs-server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "demo API"
  version: "0.0.1"
  title: "Testla Model S API"
host: "virtserver.swaggerhub.com"
basePath: "/jennyuan/Tesla-ModelS-API/0.0.1"
schemes:
- "https"
paths:
  /vehicle:
    get:
      description: "details about vehicle"
      operationId: "vehicleGET"
      parameters: []
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/inline_response_200"
      x-swagger-router-controller: "Default"
  /vehicles/{id}/data_request/drive_state:
    get:
      description: "returns driving and position state of the vehicle"
      operationId: "drive_state"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID number of the car"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/inline_response_200_1"
      x-swagger-router-controller: "Default"
  /vehicles/{id}/command/set_temps:
    post:
      description: "Set the temperature target for the HVAC system"
      operationId: "set_temps"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID number of the car"
        required: true
        type: "string"
      - name: "driver_temp"
        in: "query"
        description: "The desired temp on the driver's side\"\""
        required: true
        type: "string"
      - name: "passenger_temp"
        in: "query"
        description: "The desired temp on the passenger's side\"\""
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/inline_response_200_2"
      x-swagger-router-controller: "Default"
definitions:
  inline_response_200_1:
    properties:
      shift_state:
        type: "string"
        example: "normal"
      speed:
        type: "number"
        format: "double"
        example: 4.5
      latitude:
        type: "number"
        format: "double"
        example: 25.53844000000000136196831590496003627777099609375
    example:
      shift_state: "normal"
      latitude: 25.53844000000000136196831590496003627777099609375
      speed: 4.5
  inline_response_200:
    properties:
      color:
        type: "string"
        example: "Red"
      display_name:
        type: "string"
        example: "Jen's Model S"
      vehicle_id:
        type: "string"
        example: "ABCDE"
      vin:
        type: "string"
        example: "XYZ1234"
    example:
      color: "Red"
      vin: "XYZ1234"
      display_name: "Jen's Model S"
      vehicle_id: "ABCDE"
  inline_response_200_2:
    properties:
      result:
        type: "boolean"
        example: true
      reason:
        type: "string"
        example: "updated database"
    example:
      result: true
      reason: "updated database"